% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_generate_charts.R
\name{add_stacked_bar_tooltip}
\alias{add_stacked_bar_tooltip}
\title{Add tooltip to stacked bar chart}
\usage{
add_stacked_bar_tooltip(echart)
}
\arguments{
\item{echart}{An \code{echarts4r} object representing the stacked bar chart.}
}
\value{
An \code{echarts4r} object with the tooltip applied.
}
\description{
\code{add_stacked_bar_tooltip()} adds a tooltip to an \code{echarts4r} stacked bar chart.
The tooltip is triggered on the axis and displays values along with their percentages.
}
