% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{check_colnames}
\alias{check_colnames}
\title{Check colnames}
\usage{
check_colnames(file, expected_colnames)
}
\arguments{
\item{file}{String. Full path to a .csv file.}

\item{expected_colnames}{Character. Set of expected column names.}
}
\value{
\code{check_colnames()} does not return any value. It either produces an
error or not, so what matters is its side effect.
}
\description{
\code{check_colnames()} compares a set of expected column names with the actual
column names in a .csv file and errors if any of the expected column names
is not found in the file.
}
