% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_dm.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to PostgreSQL Database}
\usage{
connect_to_db()
}
\value{
A {DBI} database connection
}
\description{
Establish a `{DBI}` database connection to the PostgreSQL
  database, using environment variables for the connection information
}
\examples{
\dontrun{
con <- connect_to_db()
}
}
