% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CountExchangeForSexCodes}
\alias{CountExchangeForSexCodes}
\title{CountExchangeForSex Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 2 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#R15.B CountExchangeForSex>
}
\usage{
CountExchangeForSexCodes
}
\description{
CountExchangeForSex Codes
}
\keyword{datasets}
