% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_dm.R
\name{create_dm}
\alias{create_dm}
\title{Create the data model}
\usage{
create_dm()
}
\value{
List of data frames, based upon the tables in the PostgreSQL database,
  with some minor manipulations to reduce the number of data transformations
  on-the-fly in the server-side of the app
}
\description{
Create a
}
\examples{
\dontrun{
dm <- create_dm()
}
}
