% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{create_menuItems}
\alias{create_menuItems}
\title{Create menuItems based on a dataset}
\usage{
create_menuItems(menuItems)
}
\arguments{
\item{menuItems}{A tibble with columns `text` and `tabName`}
}
\value{

}
\description{
This function uses Functional Programming Tools to generate multiple elements
 with the same structure. The dataset contains one column per parameter used.
}
