% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{custom_card}
\alias{custom_card}
\title{Create a custom card}
\usage{
custom_card(..., height = "490px")
}
\arguments{
\item{...}{Additional arguments passed to \code{bslib::card()}}

\item{height}{A character string specifying the card's height. Default is \code{"490px"}.}
}
\value{
A \code{bslib::card} object.
}
\description{
\code{custom_card()} generates a Bootstrap card with a specified height and
full-screen capability.
}
