% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{custom_pickerInput}
\alias{custom_pickerInput}
\title{Create a custom pickerInput}
\usage{
custom_pickerInput(
  inputId,
  label,
  choices = NULL,
  selected = NULL,
  multiple = TRUE,
  width = "100\%",
  opts_actionsBox = TRUE,
  opts_selectedTextFormat = "count > 1",
  opts_liveSearch = TRUE,
  opts_container = "body",
  ...
)
}
\arguments{
\item{inputId}{A character string representing the input ID}

\item{label}{A character string specifying the label for the picker input}

\item{choices}{A list of choices available for selection (default: \code{NULL})}

\item{selected}{The initially selected value(s) (default: \code{NULL})}

\item{multiple}{A logical indicating whether multiple selections are allowed (default: \code{TRUE})}

\item{width}{A character string specifying the width of the input (default: \code{"100\%"})}

\item{opts_actionsBox}{A logical indicating whether to show an actions box (default: \code{TRUE})}

\item{opts_selectedTextFormat}{A character string specifying the format of the selected text (default: \code{"count > 1"})}

\item{opts_liveSearch}{A logical indicating whether to enable live search (default: \code{TRUE})}

\item{opts_container}{A character string specifying where the dropdown should be contained (default: \code{"body"})}

\item{...}{Additional arguments passed to \code{shinyWidgets::pickerInput()}.}
}
\value{
A \code{shiny.tag} object representing the picker input.
}
\description{
\code{custom_pickerInput()} is a wrapper around \code{shinyWidgets::pickerInput()} that
includes commonly used values as default.
}
