% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataCollectionStageCodes}
\alias{DataCollectionStageCodes}
\title{DataCollectionStage Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#5.03.1%20DataCollectionStage>
}
\usage{
DataCollectionStageCodes
}
\description{
DataCollectionStage Codes
}
\keyword{datasets}
