% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_process_data.R
\name{delete_from_db}
\alias{delete_from_db}
\title{Delete records from a database}
\usage{
delete_from_db(data, conn)
}
\arguments{
\item{data}{List of dataframes returned by \code{prep_tables()}.}

\item{conn}{A database connection object.}
}
\value{
Nothing. \code{delete_from_db()} is called for its side effects.
}
\description{
\code{delete_from_db()} deletes records from a database based on uploaded data.
}
\details{
\code{delete_from_db()} iterates through each table in the database (except for
'organization' and 'project') and deletes records that match on the table's
\verb{*_id} value and \code{organization_id} value, when compared to the respective
uploaded file data.
}
