% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DOBDataQualityCodes}
\alias{DOBDataQualityCodes}
\title{DOB Data Quality Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#3.03.2%20DOBDataQuality>
}
\usage{
DOBDataQualityCodes
}
\description{
DOB Data Quality Codes
}
\keyword{datasets}
