% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{filter_data}
\alias{filter_data}
\title{Filter data based on clients and validate}
\usage{
filter_data(data, clients_filtered, at = "enrollment")
}
\arguments{
\item{data}{A data frame to be filtered}

\item{clients_filtered}{A data frame containing the client records
to filter \code{data} by. Must contain columns that match the selected \code{at} level.}

\item{at}{A character string specifying the filtering level.
Either \code{"enrollment"} (default) or \code{"youth"}.
\itemize{
\item \code{"enrollment"}: Filters data based on \code{personal_id}, \code{organization_id} and \code{enrollment_id}.
\item \code{"youth"}: Filters data based on \code{personal_id} and \code{organization_id}.
}}
}
\value{
A filtered data frame containing only the records matching the specified
clients. If no matching records are found, a validation error is triggered.
}
\description{
\code{filter_data()} filters the provided dataset to include only records that match
the given \code{clients_filtered} dataset at the specified level (enrollment or youth).
It then validates the filtered dataset to ensure it is not empty.
}
