% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_process_data.R
\name{find_file}
\alias{find_file}
\title{Find the complete filepath of a certain .csv file}
\usage{
find_file(files, target)
}
\arguments{
\item{files}{Character vector containing the filepaths to search through.}

\item{target}{String specifying the name of the .csv file (without the extension)}
}
\value{
A character vector containing the filepath for the corresponding .csv file.
}
\description{
`find_file()` finds the complete filepath of a .csv file based on the .csv
basename (i.e. the file name) in a given vector of filepaths.
}
\examples{
files <- c("some/path/to/file/data1.csv", "some/path/to/file/data2.csv")
find_file(files, "data1")
}
