% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_read_data.R
\name{generate_submission_id}
\alias{generate_submission_id}
\title{Title}
\usage{
generate_submission_id(
  dir,
  contact_first_name,
  contact_last_name,
  contact_email,
  program_id,
  period_start_date,
  period_end_date
)
}
\arguments{
\item{dir}{(String) Path to the directory containing the "Submission"
.parquet file(s) in the data lake}
}
\value{

}
\description{
Title
}
