% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_process_data.R
\name{get_export_dates}
\alias{get_export_dates}
\title{Retrieve Export Dates}
\usage{
get_export_dates(dir)
}
\arguments{
\item{dir}{String, the location of the}
}
\value{

}
\description{
Retrieve Export Dates
}
