% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_generate_charts.R
\name{get_ids_for_sankey}
\alias{get_ids_for_sankey}
\title{Get eligible IDs for generating a Sankey chart}
\usage{
get_ids_for_sankey(data)
}
\arguments{
\item{data}{A data frame containing the data from which to extract unique IDs.}
}
\value{
A data frame with unique combinations of organization and personal IDs
that have at least a Project start and Project exit data collection stage.
}
\description{
`get_ids_for_sankey()` identifies IDs with both a Project start and a
Project exit data collection stage.
}
\details{
For entries with multiple Project start the most recently updated data is kept.
}
\examples{
\dontrun{
mock_data <- tibble::tribble(
  ~organization_id, ~personal_id, ~data_collection_stage, ~date_updated,
  1, 101, "Project start", "2023-01-01",
  1, 101, "Project exit", "2023-12-31",
  2, 102, "Project start", "2023-01-01",
  2, 102, "Project exit", "2023-12-31",
  2, 103, "Project start", "2023-02-01",
  3, 104, "Project start", "2023-02-01",
  3, 104, "Project update", "2023-12-15",
  3, 105, "Project start", "2023-01-01",
  3, 105, "Project update", "2023-06-30",
  3, 105, "Project exit", "2023-12-31"
)
get_ids_for_sankey(mock_data)
}
}
