% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{get_missing_categories}
\alias{get_missing_categories}
\title{Get missing data categories}
\usage{
get_missing_categories()
}
\description{
\code{get_missing_categories()} returns a character vector of standardized
categories representing missing data.
}
