% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{get_nav_panel_title}
\alias{get_nav_panel_title}
\title{Create a navigation panel title with an icon}
\usage{
get_nav_panel_title(text, icon)
}
\arguments{
\item{text}{A character string representing the title text.}

\item{icon}{A character string specifying the name of the icon from FontAwesome.}
}
\value{
A \code{shiny::span} element containing the icon and text.
}
\description{
\code{get_nav_panel_title()} generates a navigation panel title that includes an icon
and a text label displayed below it.
}
\examples{
get_nav_panel_title("Education", "book-open")
}
