% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{hash}
\alias{hash}
\title{Hash a value with a key}
\usage{
hash(x, key)
}
\arguments{
\item{x}{A character string}

\item{key}{A hashing dictionary key}
}
\value{
An hashed string
}
\description{
Hash a value with a key
}
\examples{
\dontrun{

hash("blah", key = my_key)

}
}
