% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HealthStatusCodes}
\alias{HealthStatusCodes}
\title{HealthStatus Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 2 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#R7.1 HealthStatus>
}
\usage{
HealthStatusCodes
}
\description{
HealthStatus Codes
}
\keyword{datasets}
