% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LastGradeCompletedCodes}
\alias{LastGradeCompletedCodes}
\title{LastGradeCompleted Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 2 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#R4.1%20LastGradeCompleted>
}
\usage{
LastGradeCompletedCodes
}
\description{
LastGradeCompleted Codes
}
\keyword{datasets}
