% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{link_data_standards_manual}
\alias{link_data_standards_manual}
\title{Create a Link to the HMIS Data Standards Manual}
\usage{
link_data_standards_manual()
}
\value{
An HTML anchor (\verb{<a>}) tag as a \code{shiny.tag} object.
}
\description{
\code{link_data_standards_manual()} generates an HTML anchor (\verb{<a>}) tag that links to the
HMIS Data Standards Manual PDF.
}
