% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{link_section}
\alias{link_section}
\title{Create span with hyperlink to the HMIS Data Standards Manual}
\usage{
link_section(section)
}
\arguments{
\item{section}{The name or number of the section to refer to.}
}
\value{
HTML span with a referenced section and a hyperlink to the HMIS Data Standards Manual.
}
\description{
\code{link_section()} generates a span that references a section and has a hyperlink to
the HMIS Data Standards Manual. The link opens the manual in a new browser tab.
}
\examples{
\dontrun{
link_section("R4 Last Grade Completed")
}
}
