% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LivingCodes}
\alias{LivingCodes}
\title{Living Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 3 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#3.12.1%20Living%20Situation%20Option%20List>
}
\usage{
LivingCodes
}
\description{
Living Codes
}
\keyword{datasets}
