% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MonthsHomelessPastThreeYearsCodes}
\alias{MonthsHomelessPastThreeYearsCodes}
\title{MonthsHomelessPastThreeYears Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 2 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#3.917.5%20MonthsHomelessPastThreeYears>
}
\usage{
MonthsHomelessPastThreeYearsCodes
}
\description{
MonthsHomelessPastThreeYears Codes
}
\keyword{datasets}
