% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NotEmployedReasonCodes}
\alias{NotEmployedReasonCodes}
\title{NotEmployedReason Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#R6.B NotEmployedReason>
}
\usage{
NotEmployedReasonCodes
}
\description{
NotEmployedReason Codes
}
\keyword{datasets}
