% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NoYesReasonsForMissingDataCodes}
\alias{NoYesReasonsForMissingDataCodes}
\title{NoYesReasonsForMissingData Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#1.8%20No/Yes/Reasons%20for%20Missing%20Data>
}
\usage{
NoYesReasonsForMissingDataCodes
}
\description{
NoYesReasonsForMissingData Codes
}
\keyword{datasets}
