% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_generate_charts.R
\name{pie_chart}
\alias{pie_chart}
\title{Generate a pie chart using echarts4r}
\usage{
pie_chart(data, category, count)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{category}{A character string specifying the column name in the data
frame representing the categories for the pie chart slices.}

\item{count}{A character string specifying the column name in the data frame
representing the counts or values associated with each category.}
}
\value{
A pie chart visualized using echarts4r.
}
\description{
This function generates a pie chart using the echarts4r package.
}
\examples{
mock_data <- data.frame(x = c("A", "B", "C"), y = c(10, 20, 30))
pie_chart(
  data = mock_data,
  category = "x",
  count = "y"
)
}
