% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_generate_charts.R
\name{prep_sankey_data}
\alias{prep_sankey_data}
\title{Prepare data for generating a Sankey chart}
\usage{
prep_sankey_data(data, state_var)
}
\arguments{
\item{data}{A data frame containing the data to be prepared.}

\item{state_var}{A character string specifying the column name representing
the state variable to be used in the Sankey chart.}
}
\value{
A data frame with prepared data suitable for generating a Sankey chart.
}
\description{
\code{prep_sankey_data()} prepares data by filtering, arranging, selecting, and
transforming it to generate data suitable for creating a Sankey chart
representing flow between stages.
}
\examples{
\dontrun{
mock_data <- tibble::tribble(
  ~organization_id, ~personal_id, ~data_collection_stage, ~date_updated, ~condition,
  1, 101, "Project start", "2023-01-01", "A",
  1, 101, "Project exit", "2023-12-31", "A",
  2, 102, "Project start", "2023-01-01", "A",
  2, 102, "Project exit", "2023-12-31", "B",
  2, 103, "Project start", "2023-02-01", "A",
  3, 104, "Project start", "2023-02-01", "B",
  3, 104, "Project update", "2023-12-15", "B",
  3, 105, "Project start", "2023-01-01", "B",
  3, 105, "Project update", "2023-06-30", "B",
  3, 105, "Project exit", "2023-12-31", "A"
)
prep_sankey_data(data = mock_data, state_var = "condition")
}
}
