% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_process_data.R
\name{prep_tables}
\alias{prep_tables}
\title{Prepare data for database tasks}
\usage{
prep_tables(data, conn)
}
\arguments{
\item{data}{List of dataframes returned by `process_data()`.}

\item{conn}{A database connection object.}
}
\value{
A list of processed data frames.
}
\description{
`prep_tables()` processes the list returned by `process_data()` and returns
a new list which is used by `delete_from_db()` and `send_to_db()`.
}
\details{
`prep_tables()` handles the addition of new organization and project entries to
the database, adds `project_id` and `organization_id` columns to `enrollment`
file data, adds `organization_id` to the remaining files (e.g. client,
disabilities, education, ...) and returns them as a list of dataframes.

`organization` and `project` dataframes are excluded from the returned list.
}
