% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_process_data.R
\name{process_data}
\alias{process_data}
\title{Process HMIS Data}
\usage{
process_data(file)
}
\arguments{
\item{file}{String, the full path to the .zip file containing the quarterly
HMIS data}
}
\value{

}
\description{
Process HMIS Data
}
\details{
This is the *"master"* function that governs the entire ETL process
  for processing the uploaded data and, depending on the user (i.e., whether or
  not the user is an approved grantee), writing the data out to the DropBox
  data lake
}
