% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ProjectTypeCodes}
\alias{ProjectTypeCodes}
\title{ProjectType Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 2 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#2.02.6 ProjectType>
}
\usage{
ProjectTypeCodes
}
\description{
ProjectType Codes
}
\keyword{datasets}
