% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_benefits}
\alias{read_benefits}
\title{Ingest "IncomeBenefits.csv" file and perform ETL prep for "BENEFITS"
database table}
\usage{
read_benefits(file)
}
\arguments{
\item{file}{String, the full path to the .csv file}
}
\value{
A data frame, containing the transformed data to be written out to
the "BENEFITS" database table
}
\description{
Ingest "IncomeBenefits.csv" file and perform ETL prep for "BENEFITS"
database table
}
\examples{
\dontrun{

path <- "path/to/IncomeBenefits.csv"

read_benefits(
    file = path
)
}
}
