% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_client}
\alias{read_client}
\title{Ingest "Client.csv" file and perform ETL prep for "CLIENT" database table}
\usage{
read_client(file)
}
\arguments{
\item{file}{String, the full path to the .csv file}
}
\value{
A data frame, containing the transformed data to be written out to
  the "CLIENT" database table
}
\description{
Ingest "Client.csv" file and perform ETL prep for "CLIENT" database table
}
\examples{
\dontrun{

path <- "path/to/Client.csv"

read_client(
  file = path
)

}
}
