% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_disabilities}
\alias{read_disabilities}
\title{Ingest "Disabilities.csv" file and perform ETL prep for "DISABILITIES" database table}
\usage{
read_disabilities(file)
}
\arguments{
\item{file}{String, the full path to the .csv file}
}
\value{
A data frame, containing the transformed data to be written out to
the "DISABILITIES" database table
}
\description{
Ingest "Disabilities.csv" file and perform ETL prep for "DISABILITIES" database table
}
\examples{
\dontrun{

path <- "path/to/Disabilities.csv"

read_disabilities(
  file = path,
  submission_id = 1L
)

}
}
