% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_domestic_violence}
\alias{read_domestic_violence}
\title{Ingest "HealthAndDV.csv" file and perform ETL prep for "DOMESTIC_VIOLENCE"
database table}
\usage{
read_domestic_violence(file)
}
\arguments{
\item{file}{String, the full path to the .csv file}
}
\value{
A data frame, containing the transformed data to be written out to
the "DOMESTIC_VIOLENCE" database table
}
\description{
Ingest "HealthAndDV.csv" file and perform ETL prep for "DOMESTIC_VIOLENCE"
database table
}
\examples{
\dontrun{

path <- "path/to/HealthAndDV.csv"

read_domestic_violence(
  file = path,
  submission_id = 1L
)

}
}
