% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_education}
\alias{read_education}
\title{Ingest "EmploymentEducation.csv" file and perform ETL prep for "EDUCATION" database table}
\usage{
read_education(file)
}
\arguments{
\item{file}{String, the full path to the .csv file}
}
\value{
A data frame, containing the transformed data to be written out to
the "EDUCATION" database table
}
\description{
Ingest "EmploymentEducation.csv" file and perform ETL prep for "EDUCATION" database table
}
\examples{
\dontrun{

path <- "path/to/EmploymentEducation.csv"

read_education(
    file = path
)
}
}
