% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_exit}
\alias{read_exit}
\title{Ingest "Exit.csv" file and perform ETL prep for "EXIT" database table}
\usage{
read_exit(file, submission_id)
}
\arguments{
\item{file}{String, the full path to the .csv file}

\item{submission_id}{Integer, the Submission ID associated with this upload}
}
\value{
A data frame, containing the transformed data to be written out to
  the "EXIT" database table
}
\description{
Ingest "Exit.csv" file and perform ETL prep for "EXIT" database table
}
\examples{
\dontrun{

path <- "path/to/Exit.csv"

read_exit(
  file = path,
  submission_id = 1L
)

}
}
