% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_export}
\alias{read_export}
\title{Ingest "Export.csv" file and perform ETL prep for "PROJECT" and "SUBMISSION"
database tables}
\usage{
read_export(file)
}
\arguments{
\item{file}{String, the full path to the .csv file}
}
\value{
A data frame, containing the transformed data to be written out to
the "PROJECT" and "SUBMISSION" database tables
}
\description{
Ingest "Export.csv" file and perform ETL prep for "PROJECT" and "SUBMISSION"
database tables
}
\examples{
\dontrun{

path <- "path/to/Export.csv"

read_export(
  file = path
)

}
}
