% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_funder}
\alias{read_funder}
\title{Ingest "Funder.csv" file and perform ETL prep for "FUNDER" database table}
\usage{
read_funder(file)
}
\arguments{
\item{file}{String, the full path to the .csv file}
}
\value{
A data frame, containing the transformed data to be written out to
the "FUNDER" database tables
}
\description{
Ingest "Funder.csv" file and perform ETL prep for "FUNDER" database table
}
\examples{
\dontrun{

path <- "path/to/Funder.csv"

read_funder(
  file = path
)
}
}
