% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_health}
\alias{read_health}
\title{Ingest "HealthAndDV.csv" file and perform ETL prep for "HEALTH" database
table}
\usage{
read_health(file)
}
\arguments{
\item{file}{String, the full path to the .csv file}
}
\value{
A data frame, containing the transformed data to be written out to
the "HEALTH" database table
}
\description{
Ingest "HealthAndDV.csv" file and perform ETL prep for "HEALTH" database
table
}
\examples{
\dontrun{

path <- "path/to/HealthAndDV.csv"

read_health(
    file = path
)
}
}
