% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_income}
\alias{read_income}
\title{Ingest "IncomeBenefits.csv" file and perform ETL prep for "INCOME"
database table}
\usage{
read_income(file)
}
\arguments{
\item{file}{String, the full path to the .csv file}
}
\value{
A data frame, containing the transformed data to be written out to
  the "INCOME" database table
}
\description{
Ingest "IncomeBenefits.csv" file and perform ETL prep for "INCOME"
database table
}
\examples{
\dontrun{

path <- "path/to/IncomeBenefits.csv"

read_income(
  file = path,
  submission_id = 1L
)

}
}
