% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_living}
\alias{read_living}
\title{Ingest "CurrentLivingSituation.csv" file and perform ETL prep for "LIVING"
database table}
\usage{
read_living(file)
}
\arguments{
\item{file}{String, the full path to the .csv file}
}
\value{
A data frame, containing the transformed data to be written out to
the "LIVING" database table
}
\description{
Ingest "CurrentLivingSituation.csv" file and perform ETL prep for "LIVING"
database table
}
\examples{
\dontrun{

path <- "path/to/CurrentLivingSituation.csv"

read_living(
    file = path
)
}
}
