% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_project}
\alias{read_project}
\title{Ingest "Project.csv" file and perform ETL prep for "PROGRAM" database table}
\usage{
read_project(file, submission_id)
}
\arguments{
\item{file}{String, the full path to the .csv file}

\item{ProjectID}{Integer, the Project ID associated with pulling Project information}
}
\value{
A data frame, containing the transformed data to be written out to
  the "PROGRAM" database table. "PROGRAM" database table shows ProjectName and ProjectType based on ProjectID
}
\description{
Ingest "Project.csv" file and perform ETL prep for "PROGRAM" database table
}
\examples{
\dontrun{

path <- "path/to/Project.csv"

read_project(
  file = path,
  program_id = 1L
)

}
}
