% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_project_coc}
\alias{read_project_coc}
\title{Ingest "ProjectCoC.csv" file and perform ETL prep for "PROJECT COC" database table}
\usage{
read_project_coc(file)
}
\arguments{
\item{file}{String, the full path to the .csv file}
}
\value{
A data frame, containing the transformed data to be written out to
the "PROJECT COC" database table
}
\description{
Ingest "ProjectCoC.csv" file and perform ETL prep for "PROJECT COC" database table
}
\examples{
\dontrun{

path <- "path/to/ProjectCoC.csv"

read_project_coc(
  file = path
)
}
}
