% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_services}
\alias{read_services}
\title{Ingest "Services.csv" file and perform ETL prep for "SERVICES" database table}
\usage{
read_services(file, submission_id)
}
\arguments{
\item{file}{String, the full path to the .csv file}

\item{submission_id}{Integer, the Submission ID associated with this upload}
}
\value{
A data frame, containing the transformed data to be written out to
  the "SERVICES" database table
}
\description{
Ingest "Services.csv" file and perform ETL prep for "SERVICES" database table
}
\examples{
\dontrun{

path <- "path/to/Services.csv"

read_services(
  file = path,
  submission_id = 1L
)

}
}
