% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ReferralSourceCodes}
\alias{ReferralSourceCodes}
\title{ReferralSource Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 2 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#R1.1 ReferralSource>
}
\usage{
ReferralSourceCodes
}
\description{
ReferralSource Codes
}
\keyword{datasets}
