% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RelationshipToHoHCodes}
\alias{RelationshipToHoHCodes}
\title{RelationshipToHoH Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#3.15.1%20RelationshipToHoH>
}
\usage{
RelationshipToHoHCodes
}
\description{
RelationshipToHoH Codes
}
\keyword{datasets}
