% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SchoolStatusCodes}
\alias{SchoolStatusCodes}
\title{SchoolStatus Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 2 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#R5.1 SchoolStatus>
}
\usage{
SchoolStatusCodes
}
\description{
SchoolStatus Codes
}
\keyword{datasets}
