% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SexCodes}
\alias{SexCodes}
\title{Sex Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.
}
\source{
TODO
}
\usage{
SexCodes
}
\description{
Sex Codes
}
\keyword{datasets}
