% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SexualOrientationCodes}
\alias{SexualOrientationCodes}
\title{SexualOrientation Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#R3.1 SexualOrientation>
}
\usage{
SexualOrientationCodes
}
\description{
SexualOrientation Codes
}
\keyword{datasets}
