% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{spinner_message}
\alias{spinner_message}
\title{Create message with spinner}
\usage{
spinner_message(message)
}
\arguments{
\item{message}{String. The message that should be placed below spinner}
}
\value{
HTML string that contains a spinner and text below it
}
\description{
Create message with spinner
}
