% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TimesHomelessPastThreeYearsCodes}
\alias{TimesHomelessPastThreeYearsCodes}
\title{TimesHomelessPastThreeYears Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 2 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#3.917.4%20TimesHomelessPastThreeYears>
}
\usage{
TimesHomelessPastThreeYearsCodes
}
\description{
TimesHomelessPastThreeYears Codes
}
\keyword{datasets}
