% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{validate_data}
\alias{validate_data}
\title{Validate data}
\usage{
validate_data(data, min_n = 10)
}
\arguments{
\item{data}{A data frame to validate}

\item{min_n}{An integer specifying the minimum number of rows required
to maintain confidentiality. Defaults to 10.}
}
\value{
This function does not return a value. It triggers a validation
error in a Shiny app if the dataset is empty, preventing further execution.
}
\description{
\code{validate_data()} checks whether the provided dataset meets the minimum
row requirements. It first ensures that the dataset has at least one row,
and then enforces a minimum row threshold to maintain confidentiality.
If the validation fails, a Shiny validation error is triggered, preventing
further execution.
}
