% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{validate_file_data}
\alias{validate_file_data}
\title{Validate file data}
\usage{
validate_file_data(filename, data)
}
\arguments{
\item{filename}{String. Name of the file (used in error messages).}

\item{data}{A data frame}
}
\value{
\code{validate_file_data()} is called for its side effects.
It produces an error if validation fails, otherwise nothing.
}
\description{
\code{validate_file_data()} checks whether the provided data frame is valid and
aborts with an informative error message if validation fails.
}
