% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WhenDVOccurredCodes}
\alias{WhenDVOccurredCodes}
\title{WhenDVOccurred Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 2 columns.
}
\source{
<https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf#4.11.A WhenDVOccurred>
}
\usage{
WhenDVOccurredCodes
}
\description{
WhenDVOccurred Codes
}
\keyword{datasets}
