% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{with_tooltip}
\alias{with_tooltip}
\title{Add custom Bootstrap tooltip}
\usage{
with_tooltip(text, content, placement = "right")
}
\arguments{
\item{text}{The text to display.}

\item{content}{The content to display inside the tooltip.}

\item{placement}{The placement of the tooltip relative to the text. Defaults to "right".}
}
\value{
HTML Bootstrap tooltip containing the specified content.
}
\description{
\code{with_tooltip()} creates a Bootstrap tooltip containing additional information
to display alongside a given text.
}
\examples{
\dontrun{
with_tooltip(
  text = "Some Example Text",
  content = shiny::tagList(
    shiny::span("First description line."),
    shiny::br(),
    shiny::span("Second description line.")
  )
)
}
}
