% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest_checks.R
\name{check_file_names}
\alias{check_file_names}
\title{Ensure All HMIS Data was Uploaded}
\usage{
check_file_names(dir)
}
\arguments{
\item{dir}{}
}
\value{

}
\description{
Ensure All HMIS Data was Uploaded
}
\details{
Compare the file names within the uploaded .zip file to the expected
file names (stored in the 'HMISmetadata' data object within this R package)
}
