% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_dm.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to PostgreSQL Database}
\usage{
connect_to_db(env = c("prod", "dev"))
}
\arguments{
\item{env}{Character. Indicates the database to connect to.
Valid options include \code{"prod"} and \code{"dev"}:
\itemize{
\item \code{"prod"} connects to the database in production
\item \code{"dev"} connects to an internal database that is set up using dev containers
}}
}
\value{
A {DBI} database connection
}
\description{
Establish a \code{{DBI}} database connection to the PostgreSQL
database, using environment variables for the connection information
}
\examples{
\dontrun{
con <- connect_to_db()
}
}
