% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_dm.R
\name{convert_to_ordered_factor}
\alias{convert_to_ordered_factor}
\title{Convert vector to ordered factor}
\usage{
convert_to_ordered_factor(x, codes, add_data_not_collected = FALSE)
}
\arguments{
\item{x}{A vector to be converted into an ordered factor.}

\item{codes}{A data frame containing a column \code{Description} that defines
the factor levels.}

\item{add_data_not_collected}{Logical. If \code{TRUE}, "Data not collected" is
prepended to the factor levels.}
}
\value{
An ordered factor with levels in reverse order from codes.
}
\description{
\code{convert_to_ordered_factor()} converts a vector to an ordered factor,
ordering by reverse \code{Description} from the \code{codes} data frame.
}
