% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_dm.R
\name{convert_to_ordered_factor}
\alias{convert_to_ordered_factor}
\title{Convert vector to ordered factor}
\usage{
convert_to_ordered_factor(x, codes)
}
\arguments{
\item{x}{A vector to be converted into an ordered factor.}

\item{codes}{A data frame containing a column \code{Description} that defines
the factor levels (excluding "Missing").}
}
\value{
An ordered factor with levels starting with "Missing" and followed
by the reversed descriptions.
}
\description{
\code{convert_to_ordered_factor()} converts a vector to an ordered factor,
ensuring that "Missing" appears as the first level, followed by the
descriptions from a provided \code{codes} data frame in reverse order.
}
