% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CountyCodes}
\alias{CountyCodes}
\title{CountyCodes Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 123 rows and 2 columns.
}
\source{
\url{https://www.hud.gov/sites/dfiles/CPD/documents/CoC/FY2024-GeoCodes-Report.pdf}
}
\usage{
CountyCodes
}
\description{
A county can contain multiple geocodes.
}
\keyword{datasets}
