% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{link_section}
\alias{link_section}
\title{Create hyperlink to specific section in the HMIS Data Standards Manual}
\usage{
link_section(section, label = "HMIS Data Standards Manual")
}
\arguments{
\item{section}{The name or number of the section to link to.}

\item{label}{The text to display as the link. Defaults to "HMIS Data Standards Manual".}
}
\value{
HTML hyperlink to the specified section in the HMIS Data Standards Manual.
}
\description{
This function generates a hyperlink to a specific section within the HMIS Data
Standards Manual. The link opens the manual in a new browser tab.
}
\examples{
\dontrun{
link_section("R4 Last Grade Completed")
}
}
