% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_client}
\alias{read_client}
\title{Ingest "Client.csv" file and perform ETL prep for data lake}
\usage{
read_client(file, submission_id)
}
\arguments{
\item{file}{String, the full path to the .csv file}

\item{submission_id}{Integer, the Submission ID associated with this upload}
}
\value{
A data frame, containing the transformed data to be written out to
  the data lake as a .parquet file
}
\description{
Ingest "Client.csv" file and perform ETL prep for data lake
}
\examples{
\dontrun{

path <- "path/to/Client.csv"

read_client(
  file = path,
  submission_id = 1L
)

}
}
