% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_enrollment}
\alias{read_enrollment}
\title{Ingest "Enrollment.csv" file and perform ETL prep for "ENROLLMENT" database
table}
\usage{
read_enrollment(file)
}
\arguments{
\item{file}{String, the full path to the .csv file}
}
\value{
A data frame, containing the transformed data to be written out to
  the "ENROLLMENT" database table
}
\description{
Ingest "Enrollment.csv" file and perform ETL prep for "ENROLLMENT" database
table
}
\examples{
\dontrun{

path <- "path/to/Enrollment.csv"

read_enrollment(
  file = path,
  submission_id = 1L
)

}
}
