% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ingest.R
\name{read_project}
\alias{read_project}
\title{Ingest "Project.csv" file and perform ETL prep for "PROJECT" database table}
\usage{
read_project(file)
}
\arguments{
\item{file}{String, the full path to the .csv file}
}
\value{
A data frame, containing the transformed data to be written out to
the "PROJECT" database table
}
\description{
Ingest "Project.csv" file and perform ETL prep for "PROJECT" database table
}
\examples{
\dontrun{

path <- "path/to/Project.csv"

read_project(
  file = path
)

}
}
