% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{validate_data}
\alias{validate_data}
\title{Validate data}
\usage{
validate_data(data, message = "No data to display")
}
\arguments{
\item{data}{A data frame to validate}

\item{message}{A character string specifying the message to display when
the data is empty. Defaults to "No data to display".}
}
\value{
This function does not return a value. It triggers a validation
error in a Shiny app if the dataset is empty, preventing further execution.
}
\description{
\code{validate_data()} checks whether the provided dataset has at least one row.
If the data is empty, it triggers a validation error with a specified message.
}
