% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{with_popover}
\alias{with_popover}
\title{Add custom Bootstrap popover}
\usage{
with_popover(text, content, placement = "right")
}
\arguments{
\item{text}{The text to display.}

\item{content}{The content to display inside the popover.}

\item{placement}{The placement of the popover relative to the text. Defaults to "right".}
}
\value{
HTML Bootstrap popover containing the specified content.
}
\description{
\code{with_popover()} creates a Bootstrap popover containing additional information
to display alongside a given text.
}
\examples{
\dontrun{
with_popover(
  text = "Some Example Text",
  content = shiny::tagList(
    shiny::span("First description line."),
    shiny::br(),
    shiny::span("Second description line.")
  )
)
}
}
